refactor(measured-boot): move grpc conversions into rpc crate
Move measured-boot protobuf conversion traits and impls out of the
measured-boot crate and into the rpc crate. This removes the transitional
measured-boot -> rpc dependency and eliminates a transitive rpc dependency
from downstream database/model crates.

This is part of the broader effort to remove direct model -> rpc and db -> rpc
dependencies by keeping grpc-specific conversion logic alongside the generated
protobuf types.

Also move measured-boot CLI table/summary helpers into the measured-boot crate
so admin-cli output formatting no longer depends on rpc for those traits.
