+++ crates/admin-cli/src/attestation/measured_boot/report/cmds.rs
@@ -18,9 +18,9 @@
 //!
 //! `measurement report` subcommand dispatcher + backing functions.
 
-use ::rpc::admin_cli::ToTable;
+use ::rpc::measured_boot::FromGrpcOpt;
 use ::rpc::protos::measured_boot::ListMeasurementReportRequest;
-use measured_boot::FromGrpcOpt;
+use measured_boot::ToTable;
 use measured_boot::bundle::MeasurementBundle;
 use measured_boot::records::MeasurementReportRecord;
 use measured_boot::report::MeasurementReport;
