+++ crates/measured-boot/Cargo.toml
@@ -26,12 +26,11 @@ name = "measured_boot"
 
 [features]
 sqlx = ["dep:sqlx", "carbide-uuid/sqlx"]
-cli = ["dep:prettytable-rs", "dep:eyre", "dep:clap", "carbide-rpc/cli"]
+cli = ["dep:prettytable-rs", "dep:eyre", "dep:clap"]
 
 [dependencies]
 # workspace dependencies
 carbide-uuid = { path = "../uuid" }
-carbide-rpc = { path = "../rpc" }
 
 thiserror = { workspace = true }
 sqlx = { workspace = true, optional = true }
