+++ crates/measured-boot/src/bundle.rs
@@ -21,15 +21,12 @@
  */
 
 use carbide_uuid::measured_boot::{MeasurementBundleId, MeasurementSystemProfileId};
-#[cfg(feature = "cli")]
-use rpc::admin_cli::ToTable;
-use rpc::errors::RpcDataConversionError;
-use rpc::protos::measured_boot::{MeasurementBundlePb, MeasurementBundleStatePb};
 use serde::Serialize;
 
 use super::pcr::PcrRegisterValue;
 use super::records::{MeasurementBundleState, MeasurementBundleValueRecord};
-use crate::{FromGrpc, FromGrpcOpt};
+#[cfg(feature = "cli")]
+use crate::ToTable;
 
 /// MeasurementBundle is a composition of a MeasurementBundleRecord,
 /// whose attributes are essentially copied directly it, as well as
@@ -66,24 +63,6 @@ pub struct MeasurementBundle {
     pub ts: chrono::DateTime<chrono::Utc>,
 }
 
-impl From<MeasurementBundle> for MeasurementBundlePb {
-    fn from(val: MeasurementBundle) -> Self {
-        let pb_state: MeasurementBundleStatePb = val.state.into();
-        Self {
-            bundle_id: Some(val.bundle_id),
-            profile_id: Some(val.profile_id),
-            name: val.name,
-            state: pb_state.into(),
-            values: val
-                .values
-                .iter()
-                .map(|value| value.clone().into())
-                .collect(),
-            ts: Some(val.ts.into()),
-        }
-    }
-}
-
 impl MeasurementBundle {
     pub fn pcr_values(&self) -> Vec<PcrRegisterValue> {
         let borrowed = &self.values;
@@ -94,43 +73,6 @@ impl MeasurementBundle {
 impl crate::DisplayName for MeasurementBundle {
     fn display_name() -> &'static str {
         "bundle"
-    }
-}
-
-impl FromGrpc<MeasurementBundlePb> for MeasurementBundle {}
-
-impl FromGrpcOpt<MeasurementBundlePb> for MeasurementBundle {}
-
-impl TryFrom<MeasurementBundlePb> for MeasurementBundle {
-    type Error = Box<dyn std::error::Error>;
-
-    fn try_from(msg: MeasurementBundlePb) -> Result<Self, Box<dyn std::error::Error>> {
-        let state = msg.state();
-        let values: super::Result<Vec<MeasurementBundleValueRecord>> = msg
-            .values
-            .iter()
-            .map(
-                |attr| match MeasurementBundleValueRecord::try_from(attr.clone()) {
-                    Ok(worked) => Ok(worked),
-                    Err(failed) => Err(super::Error::RpcConversion(format!(
-                        "attr conversion failed: {failed}"
-                    ))),
-                },
-            )
-            .collect();
-
-        Ok(Self {
-            bundle_id: msg
-                .bundle_id
-                .ok_or(RpcDataConversionError::MissingArgument("bundle_id"))?,
-            profile_id: msg
-                .profile_id
-                .ok_or(RpcDataConversionError::MissingArgument("profile_id"))?,
-            name: msg.name.clone(),
-            state: MeasurementBundleState::from(state),
-            values: values?,
-            ts: chrono::DateTime::<chrono::Utc>::try_from(msg.ts.unwrap())?,
-        })
     }
 }
