+++ crates/measured-boot/src/journal.rs
@@ -20,25 +20,19 @@
  *  tables in the database, leveraging the journal-specific record types.
  */
 
-use std::str::FromStr;
-
-use carbide_uuid::UuidEmptyStringError;
 use carbide_uuid::machine::MachineId;
 use carbide_uuid::measured_boot::{
     MeasurementBundleId, MeasurementJournalId, MeasurementReportId, MeasurementSystemProfileId,
 };
 use chrono::Utc;
-use rpc::errors::RpcDataConversionError;
-use rpc::protos::measured_boot::{MeasurementJournalPb, MeasurementMachineStatePb};
 use serde::Serialize;
 #[cfg(feature = "cli")]
 use {
-    rpc::admin_cli::ToTable,
-    rpc::admin_cli::{just_print_summary, serde_just_print_summary},
+    crate::ToTable,
+    crate::{just_print_summary, serde_just_print_summary},
 };
 
 use super::records::MeasurementMachineState;
-use crate::{FromGrpc, FromGrpcOpt};
 
 /// MeasurementJournal is a composition of a MeasurementJournalRecord,
 /// whose attributes are essentially copied directly it, as well as
@@ -92,50 +86,6 @@ impl MeasurementJournal {
 impl crate::DisplayName for MeasurementJournal {
     fn display_name() -> &'static str {
         "journal"
-    }
-}
-
-impl FromGrpc<MeasurementJournalPb> for MeasurementJournal {}
-
-impl FromGrpcOpt<MeasurementJournalPb> for MeasurementJournal {}
-
-impl From<MeasurementJournal> for MeasurementJournalPb {
-    fn from(val: MeasurementJournal) -> Self {
-        let pb_state: MeasurementMachineStatePb = val.state.into();
-        Self {
-            journal_id: Some(val.journal_id),
-            machine_id: val.machine_id.to_string(),
-            report_id: Some(val.report_id),
-            profile_id: val.profile_id,
-            bundle_id: val.bundle_id,
-            state: pb_state.into(),
-            ts: Some(val.ts.into()),
-        }
-    }
-}
-
-impl TryFrom<MeasurementJournalPb> for MeasurementJournal {
-    type Error = Box<dyn std::error::Error>;
-
-    fn try_from(msg: MeasurementJournalPb) -> Result<Self, Box<dyn std::error::Error>> {
-        if msg.machine_id.is_empty() {
-            return Err(UuidEmptyStringError {}.into());
-        }
-        let state = msg.state();
-
-        Ok(Self {
-            journal_id: msg
-                .journal_id
-                .ok_or(RpcDataConversionError::MissingArgument("journal_id"))?,
-            machine_id: MachineId::from_str(&msg.machine_id)?,
-            report_id: msg
-                .report_id
-                .ok_or(RpcDataConversionError::MissingArgument("report_id"))?,
-            profile_id: msg.profile_id,
-            bundle_id: msg.bundle_id,
-            state: MeasurementMachineState::from(state),
-            ts: chrono::DateTime::<chrono::Utc>::try_from(msg.ts.unwrap())?,
-        })
     }
 }
