+++ crates/measured-boot/src/records.rs
@@ -27,29 +27,18 @@
  *  what type of key is being passed around. A bunch of uuid::Uuid is meh.
  */
 
-use std::convert::Into;
 use std::error::Error;
 use std::fmt;
 use std::str::FromStr;
 
+use carbide_uuid::DbTable;
 use carbide_uuid::machine::MachineId;
 use carbide_uuid::measured_boot::{
     MeasurementApprovedMachineId, MeasurementApprovedProfileId, MeasurementBundleId,
     MeasurementBundleValueId, MeasurementJournalId, MeasurementReportId, MeasurementReportValueId,
     MeasurementSystemProfileAttrId, MeasurementSystemProfileId, TrustedMachineId,
 };
-use carbide_uuid::{DbTable, UuidEmptyStringError};
 use chrono::{DateTime, Utc};
-#[cfg(feature = "cli")]
-use rpc::admin_cli::{ToTable, serde_just_print_summary};
-use rpc::errors::RpcDataConversionError;
-use rpc::protos::measured_boot::{
-    CandidateMachineSummaryPb, MeasurementApprovedMachineRecordPb,
-    MeasurementApprovedProfileRecordPb, MeasurementApprovedTypePb, MeasurementBundleRecordPb,
-    MeasurementBundleStatePb, MeasurementBundleValueRecordPb, MeasurementJournalRecordPb,
-    MeasurementMachineStatePb, MeasurementReportRecordPb, MeasurementReportValueRecordPb,
-    MeasurementSystemProfileAttrRecordPb, MeasurementSystemProfileRecordPb,
-};
 use serde::{Deserialize, Serialize};
 #[cfg(feature = "sqlx")]
 use sqlx::{
@@ -57,8 +46,9 @@ use sqlx::{
     {FromRow, Row},
 };
 
-use super::pcr::PcrRegisterValue;
-use crate::{DisplayName, FromGrpc, FromGrpcOpt, FromPbVec};
+use crate::DisplayName;
+#[cfg(feature = "cli")]
+use crate::{ToTable, serde_just_print_summary};
 
 /// StringToEnumError is used for taking an input string and converting
 /// it to an enum of a given type. It is leveraged by MeasurementBundleState,
@@ -94,7 +84,7 @@ pub struct MeasurementSystemProfileRecord {
     pub name: String,
 
     // ts is the timestamp the profile was created.
-    pub ts: chrono::DateTime<Utc>,
+    pub ts: DateTime<Utc>,
 }
 
 impl DbTable for MeasurementSystemProfileRecord {
@@ -106,34 +96,6 @@ impl DbTable for MeasurementSystemProfileRecord {
 impl DisplayName for MeasurementSystemProfileRecord {
     fn display_name() -> &'static str {
         "system profile record"
-    }
-}
-
-impl FromGrpc<MeasurementSystemProfileRecordPb> for MeasurementSystemProfileRecord {}
-
-impl FromPbVec<MeasurementSystemProfileRecordPb> for MeasurementSystemProfileRecord {}
-
-impl From<MeasurementSystemProfileRecord> for MeasurementSystemProfileRecordPb {
-    fn from(val: MeasurementSystemProfileRecord) -> Self {
-        Self {
-            profile_id: Some(val.profile_id),
-            name: val.name,
-            ts: Some(val.ts.into()),
-        }
-    }
-}
-
-impl TryFrom<MeasurementSystemProfileRecordPb> for MeasurementSystemProfileRecord {
-    type Error = Box<dyn std::error::Error>;
-
-    fn try_from(msg: MeasurementSystemProfileRecordPb) -> Result<Self, Box<dyn std::error::Error>> {
-        Ok(Self {
-            profile_id: msg
-                .profile_id
-                .ok_or(RpcDataConversionError::MissingArgument("profile_id"))?,
-            name: msg.name.clone(),
-            ts: DateTime::<Utc>::try_from(msg.ts.unwrap())?,
-        })
     }
 }
 
@@ -186,42 +148,6 @@ impl DbTable for MeasurementSystemProfileAttrRecord {
 impl DisplayName for MeasurementSystemProfileAttrRecord {
     fn display_name() -> &'static str {
         "system profile attr record"
-    }
-}
-
-impl FromGrpc<MeasurementSystemProfileAttrRecordPb> for MeasurementSystemProfileAttrRecord {}
-
-impl FromPbVec<MeasurementSystemProfileAttrRecordPb> for MeasurementSystemProfileAttrRecord {}
-
-impl From<MeasurementSystemProfileAttrRecord> for MeasurementSystemProfileAttrRecordPb {
-    fn from(val: MeasurementSystemProfileAttrRecord) -> Self {
-        Self {
-            attribute_id: Some(val.attribute_id),
-            profile_id: Some(val.profile_id),
-            key: val.key,
-            value: val.value,
-            ts: Some(val.ts.into()),
-        }
-    }
-}
-
-impl TryFrom<MeasurementSystemProfileAttrRecordPb> for MeasurementSystemProfileAttrRecord {
-    type Error = Box<dyn std::error::Error>;
-
-    fn try_from(
-        msg: MeasurementSystemProfileAttrRecordPb,
-    ) -> Result<Self, Box<dyn std::error::Error>> {
-        Ok(Self {
-            attribute_id: msg
-                .attribute_id
-                .ok_or(RpcDataConversionError::MissingArgument("attribute_id"))?,
-            profile_id: msg
-                .profile_id
-                .ok_or(RpcDataConversionError::MissingArgument("profile_id"))?,
-            key: msg.key.clone(),
-            value: msg.value.clone(),
-            ts: DateTime::<Utc>::try_from(msg.ts.unwrap())?,
-        })
     }
 }
 
@@ -293,30 +219,6 @@ impl fmt::Display for MeasurementBundleState {
     }
 }
 
-impl From<MeasurementBundleState> for MeasurementBundleStatePb {
-    fn from(val: MeasurementBundleState) -> Self {
-        match val {
-            MeasurementBundleState::Pending => Self::Pending,
-            MeasurementBundleState::Active => Self::Active,
-            MeasurementBundleState::Obsolete => Self::Obsolete,
-            MeasurementBundleState::Retired => Self::Retired,
-            MeasurementBundleState::Revoked => Self::Revoked,
-        }
-    }
-}
-
-impl From<MeasurementBundleStatePb> for MeasurementBundleState {
-    fn from(msg: MeasurementBundleStatePb) -> Self {
-        match msg {
-            MeasurementBundleStatePb::Pending => Self::Pending,
-            MeasurementBundleStatePb::Active => Self::Active,
-            MeasurementBundleStatePb::Obsolete => Self::Obsolete,
-            MeasurementBundleStatePb::Retired => Self::Retired,
-            MeasurementBundleStatePb::Revoked => Self::Revoked,
-        }
-    }
-}
-
 /// MeasurementBundleStateRecord exists so we can do an sqlx::query_as and
 /// *just* select the state (and bind it to a struct). It doesn't really need
 /// to be much other than this for now.
@@ -371,43 +273,6 @@ impl DisplayName for MeasurementBundleRecord {
     }
 }
 
-impl FromGrpc<MeasurementBundleRecordPb> for MeasurementBundleRecord {}
-
-impl FromPbVec<MeasurementBundleRecordPb> for MeasurementBundleRecord {}
-
-impl From<MeasurementBundleRecord> for MeasurementBundleRecordPb {
-    fn from(val: MeasurementBundleRecord) -> Self {
-        let pb_state: MeasurementBundleStatePb = val.state.into();
-        Self {
-            bundle_id: Some(val.bundle_id),
-            name: val.name,
-            profile_id: Some(val.profile_id),
-            state: pb_state.into(),
-            ts: Some(val.ts.into()),
-        }
-    }
-}
-
-impl TryFrom<MeasurementBundleRecordPb> for MeasurementBundleRecord {
-    type Error = Box<dyn std::error::Error>;
-
-    fn try_from(msg: MeasurementBundleRecordPb) -> Result<Self, Box<dyn std::error::Error>> {
-        let state = msg.state();
-
-        Ok(Self {
-            bundle_id: msg
-                .bundle_id
-                .ok_or(RpcDataConversionError::MissingArgument("bundle_id"))?,
-            profile_id: msg
-                .profile_id
-                .ok_or(RpcDataConversionError::MissingArgument("profile_id"))?,
-            name: msg.name.clone(),
-            state: MeasurementBundleState::from(state),
-            ts: DateTime::<Utc>::try_from(msg.ts.unwrap())?,
-        })
-    }
-}
-
 /// MeasurementBundleValueRecord defines a single row
 /// from the measurement_bundles_values table.
 ///
@@ -457,49 +322,6 @@ impl DisplayName for MeasurementBundleValueRecord {
     }
 }
 
-impl FromGrpc<MeasurementBundleValueRecordPb> for MeasurementBundleValueRecord {}
-
-impl FromPbVec<MeasurementBundleValueRecordPb> for MeasurementBundleValueRecord {}
-
-impl From<MeasurementBundleValueRecord> for PcrRegisterValue {
-    fn from(val: MeasurementBundleValueRecord) -> Self {
-        PcrRegisterValue {
-            pcr_register: val.pcr_register,
-            sha_any: val.sha_any,
-        }
-    }
-}
-
-impl From<MeasurementBundleValueRecord> for MeasurementBundleValueRecordPb {
-    fn from(val: MeasurementBundleValueRecord) -> Self {
-        Self {
-            value_id: Some(val.value_id),
-            bundle_id: Some(val.bundle_id),
-            pcr_register: val.pcr_register as i32,
-            sha_any: val.sha_any,
-            ts: Some(val.ts.into()),
-        }
-    }
-}
-
-impl TryFrom<MeasurementBundleValueRecordPb> for MeasurementBundleValueRecord {
-    type Error = Box<dyn std::error::Error>;
-
-    fn try_from(msg: MeasurementBundleValueRecordPb) -> Result<Self, Box<dyn std::error::Error>> {
-        Ok(Self {
-            value_id: msg
-                .value_id
-                .ok_or(RpcDataConversionError::MissingArgument("value_id"))?,
-            bundle_id: msg
-                .bundle_id
-                .ok_or(RpcDataConversionError::MissingArgument("bundle_id"))?,
-            pcr_register: msg.pcr_register as i16,
-            sha_any: msg.sha_any.clone(),
-            ts: DateTime::<Utc>::try_from(msg.ts.unwrap())?,
-        })
-    }
-}
-
 /// MeasurementReportRecord defines a single row from
 /// the measurement_reports table.
 ///
@@ -527,36 +349,6 @@ impl DbTable for MeasurementReportRecord {
 impl DisplayName for MeasurementReportRecord {
     fn display_name() -> &'static str {
         "report record"
-    }
-}
-
-impl FromGrpc<MeasurementReportRecordPb> for MeasurementReportRecord {}
-
-impl From<MeasurementReportRecord> for MeasurementReportRecordPb {
-    fn from(val: MeasurementReportRecord) -> Self {
-        Self {
-            report_id: Some(val.report_id),
-            machine_id: val.machine_id.to_string(),
-            ts: Some(val.ts.into()),
-        }
-    }
-}
-
-impl TryFrom<MeasurementReportRecordPb> for MeasurementReportRecord {
-    type Error = Box<dyn std::error::Error>;
-
-    fn try_from(msg: MeasurementReportRecordPb) -> Result<Self, Box<dyn std::error::Error>> {
-        if msg.machine_id.is_empty() {
-            return Err(UuidEmptyStringError {}.into());
-        };
-
-        Ok(Self {
-            report_id: msg
-                .report_id
-                .ok_or(RpcDataConversionError::MissingArgument("report_id"))?,
-            machine_id: MachineId::from_str(&msg.machine_id)?,
-            ts: DateTime::<Utc>::try_from(msg.ts.unwrap())?,
-        })
     }
 }
 
@@ -608,47 +400,6 @@ impl DbTable for MeasurementReportValueRecord {
 impl DisplayName for MeasurementReportValueRecord {
     fn display_name() -> &'static str {
         "report value record"
-    }
-}
-
-impl FromGrpc<MeasurementReportValueRecordPb> for MeasurementReportValueRecord {}
-
-impl From<MeasurementReportValueRecord> for PcrRegisterValue {
-    fn from(val: MeasurementReportValueRecord) -> Self {
-        Self {
-            pcr_register: val.pcr_register,
-            sha_any: val.sha_any,
-        }
-    }
-}
-
-impl From<MeasurementReportValueRecord> for MeasurementReportValueRecordPb {
-    fn from(val: MeasurementReportValueRecord) -> Self {
-        Self {
-            value_id: Some(val.value_id),
-            report_id: Some(val.report_id),
-            pcr_register: val.pcr_register as i32,
-            sha_any: val.sha_any,
-            ts: Some(val.ts.into()),
-        }
-    }
-}
-
-impl TryFrom<MeasurementReportValueRecordPb> for MeasurementReportValueRecord {
-    type Error = Box<dyn std::error::Error>;
-
-    fn try_from(msg: MeasurementReportValueRecordPb) -> Result<Self, Box<dyn std::error::Error>> {
-        Ok(Self {
-            value_id: msg
-                .value_id
-                .ok_or(RpcDataConversionError::MissingArgument("value_id"))?,
-            report_id: msg
-                .report_id
-                .ok_or(RpcDataConversionError::MissingArgument("report_id"))?,
-            pcr_register: msg.pcr_register as i16,
-            sha_any: msg.sha_any.clone(),
-            ts: DateTime::<Utc>::try_from(msg.ts.unwrap())?,
-        })
     }
 }
 
@@ -714,49 +465,6 @@ impl DisplayName for MeasurementJournalRecord {
     }
 }
 
-impl FromGrpc<MeasurementJournalRecordPb> for MeasurementJournalRecord {}
-
-impl From<MeasurementJournalRecord> for MeasurementJournalRecordPb {
-    fn from(val: MeasurementJournalRecord) -> Self {
-        let pb_state: MeasurementMachineStatePb = val.state.into();
-
-        Self {
-            journal_id: Some(val.journal_id),
-            machine_id: val.machine_id.to_string(),
-            report_id: Some(val.report_id),
-            profile_id: val.profile_id,
-            bundle_id: val.bundle_id,
-            state: pb_state.into(),
-            ts: Some(val.ts.into()),
-        }
-    }
-}
-
-impl TryFrom<MeasurementJournalRecordPb> for MeasurementJournalRecord {
-    type Error = Box<dyn std::error::Error>;
-
-    fn try_from(msg: MeasurementJournalRecordPb) -> Result<Self, Box<dyn std::error::Error>> {
-        if msg.machine_id.is_empty() {
-            return Err(UuidEmptyStringError {}.into());
-        }
-        let state = msg.state();
-
-        Ok(Self {
-            journal_id: msg
-                .journal_id
-                .ok_or(RpcDataConversionError::MissingArgument("journal_id"))?,
-            machine_id: MachineId::from_str(&msg.machine_id)?,
-            report_id: msg
-                .report_id
-                .ok_or(RpcDataConversionError::MissingArgument("report_id"))?,
-            profile_id: msg.profile_id,
-            bundle_id: msg.bundle_id,
-            state: MeasurementMachineState::from(state),
-            ts: DateTime::<Utc>::try_from(msg.ts.unwrap())?,
-        })
-    }
-}
-
 /// MeasurementMachineState is an enum in the database, and
 /// is used for tracking the state of a machine.
 ///
@@ -795,28 +503,6 @@ impl fmt::Display for MeasurementMachineState {
     }
 }
 
-impl From<MeasurementMachineState> for MeasurementMachineStatePb {
-    fn from(val: MeasurementMachineState) -> Self {
-        match val {
-            MeasurementMachineState::Discovered => Self::Discovered,
-            MeasurementMachineState::PendingBundle => Self::PendingBundle,
-            MeasurementMachineState::Measured => Self::Measured,
-            MeasurementMachineState::MeasuringFailed => Self::MeasuringFailed,
-        }
-    }
-}
-
-impl From<MeasurementMachineStatePb> for MeasurementMachineState {
-    fn from(msg: MeasurementMachineStatePb) -> Self {
-        match msg {
-            MeasurementMachineStatePb::Discovered => Self::Discovered,
-            MeasurementMachineStatePb::PendingBundle => Self::PendingBundle,
-            MeasurementMachineStatePb::Measured => Self::Measured,
-            MeasurementMachineStatePb::MeasuringFailed => Self::MeasuringFailed,
-        }
-    }
-}
-
 #[derive(Debug, Clone, Serialize)]
 pub struct CandidateMachineSummary {
     // machine_id is the ID of the machine, e.g. fm100hxxxxx.
@@ -829,32 +515,6 @@ pub struct CandidateMachineSummary {
 impl DisplayName for CandidateMachineSummary {
     fn display_name() -> &'static str {
         "candidate machine record"
-    }
-}
-
-impl FromGrpc<CandidateMachineSummaryPb> for CandidateMachineSummary {}
-
-impl From<CandidateMachineSummary> for CandidateMachineSummaryPb {
-    fn from(val: CandidateMachineSummary) -> Self {
-        Self {
-            machine_id: val.machine_id.to_string(),
-            ts: Some(val.ts.into()),
-        }
-    }
-}
-
-impl TryFrom<CandidateMachineSummaryPb> for CandidateMachineSummary {
-    type Error = Box<dyn std::error::Error>;
-
-    fn try_from(msg: CandidateMachineSummaryPb) -> Result<Self, Box<dyn std::error::Error>> {
-        if msg.machine_id.is_empty() {
-            return Err(UuidEmptyStringError {}.into());
-        }
-
-        Ok(Self {
-            machine_id: MachineId::from_str(&msg.machine_id)?,
-            ts: DateTime::<Utc>::try_from(msg.ts.unwrap())?,
-        })
     }
 }
 
@@ -900,24 +560,6 @@ impl FromStr for MeasurementApprovedType {
 impl fmt::Display for MeasurementApprovedType {
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
         write!(f, "{self:?}")
-    }
-}
-
-impl From<MeasurementApprovedType> for MeasurementApprovedTypePb {
-    fn from(val: MeasurementApprovedType) -> Self {
-        match val {
-            MeasurementApprovedType::Oneshot => Self::Oneshot,
-            MeasurementApprovedType::Persist => Self::Persist,
-        }
-    }
-}
-
-impl From<MeasurementApprovedTypePb> for MeasurementApprovedType {
-    fn from(val: MeasurementApprovedTypePb) -> Self {
-        match val {
-            MeasurementApprovedTypePb::Oneshot => Self::Oneshot,
-            MeasurementApprovedTypePb::Persist => Self::Persist,
-        }
     }
 }
 
@@ -976,58 +618,9 @@ impl DisplayName for MeasurementApprovedMachineRecord {
     }
 }
 
-impl FromGrpc<MeasurementApprovedMachineRecordPb> for MeasurementApprovedMachineRecord {}
-
-impl FromGrpcOpt<MeasurementApprovedMachineRecordPb> for MeasurementApprovedMachineRecord {}
-
 impl DbTable for MeasurementApprovedMachineRecord {
     fn db_table_name() -> &'static str {
         "measurement_approved_machines"
-    }
-}
-
-impl From<MeasurementApprovedMachineRecord> for MeasurementApprovedMachineRecordPb {
-    fn from(val: MeasurementApprovedMachineRecord) -> Self {
-        let approval_type: MeasurementApprovedTypePb = val.approval_type.into();
-
-        Self {
-            approval_id: Some(val.approval_id),
-            machine_id: val.machine_id.to_string(),
-            approval_type: approval_type.into(),
-            pcr_registers: val.pcr_registers.unwrap_or("".to_string()),
-            comments: val.comments.unwrap_or("".to_string()),
-            ts: Some(val.ts.into()),
-        }
-    }
-}
-
-impl TryFrom<MeasurementApprovedMachineRecordPb> for MeasurementApprovedMachineRecord {
-    type Error = Box<dyn std::error::Error>;
-
-    fn try_from(
-        msg: MeasurementApprovedMachineRecordPb,
-    ) -> Result<Self, Box<dyn std::error::Error>> {
-        if msg.machine_id.is_empty() {
-            return Err(UuidEmptyStringError {}.into());
-        }
-        let approval_type = msg.approval_type();
-
-        Ok(Self {
-            approval_id: msg
-                .approval_id
-                .ok_or(RpcDataConversionError::MissingArgument("approval_id"))?,
-            machine_id: TrustedMachineId::from_str(&msg.machine_id)?,
-            approval_type: MeasurementApprovedType::from(approval_type),
-            pcr_registers: match !msg.pcr_registers.is_empty() {
-                true => Some(msg.pcr_registers.clone()),
-                false => None,
-            },
-            comments: match !msg.comments.is_empty() {
-                true => Some(msg.comments.clone()),
-                false => None,
-            },
-            ts: DateTime::<Utc>::try_from(msg.ts.unwrap())?,
-        })
     }
 }
 
@@ -1084,62 +677,15 @@ pub struct MeasurementApprovedProfileRecord {
     pub ts: chrono::DateTime<Utc>,
 }
 
-impl DisplayName for MeasurementApprovedProfileRecord {
-    fn display_name() -> &'static str {
-        "record"
-    }
-}
-
-impl FromGrpc<MeasurementApprovedProfileRecordPb> for MeasurementApprovedProfileRecord {}
-
-impl FromGrpcOpt<MeasurementApprovedProfileRecordPb> for MeasurementApprovedProfileRecord {}
-
 impl DbTable for MeasurementApprovedProfileRecord {
     fn db_table_name() -> &'static str {
         "measurement_approved_profiles"
     }
 }
 
-impl From<MeasurementApprovedProfileRecord> for MeasurementApprovedProfileRecordPb {
-    fn from(val: MeasurementApprovedProfileRecord) -> Self {
-        let approval_type: MeasurementApprovedTypePb = val.approval_type.into();
-
-        Self {
-            approval_id: Some(val.approval_id),
-            profile_id: Some(val.profile_id),
-            approval_type: approval_type.into(),
-            pcr_registers: val.pcr_registers.unwrap_or("".to_string()),
-            comments: val.comments.unwrap_or("".to_string()),
-            ts: Some(val.ts.into()),
-        }
-    }
-}
-
-impl TryFrom<MeasurementApprovedProfileRecordPb> for MeasurementApprovedProfileRecord {
-    type Error = Box<dyn std::error::Error>;
-
-    fn try_from(
-        msg: MeasurementApprovedProfileRecordPb,
-    ) -> Result<Self, Box<dyn std::error::Error>> {
-        let approval_type = msg.approval_type();
-        Ok(Self {
-            approval_id: msg
-                .approval_id
-                .ok_or(RpcDataConversionError::MissingArgument("approval_id"))?,
-            profile_id: msg
-                .profile_id
-                .ok_or(RpcDataConversionError::MissingArgument("profile_id"))?,
-            approval_type: MeasurementApprovedType::from(approval_type),
-            pcr_registers: match !msg.pcr_registers.is_empty() {
-                true => Some(msg.pcr_registers.clone()),
-                false => None,
-            },
-            comments: match !msg.comments.is_empty() {
-                true => Some(msg.comments.clone()),
-                false => None,
-            },
-            ts: DateTime::<Utc>::try_from(msg.ts.unwrap())?,
-        })
+impl DisplayName for MeasurementApprovedProfileRecord {
+    fn display_name() -> &'static str {
+        "record"
     }
 }
