+++ crates/measured-boot/src/report.rs
@@ -20,21 +20,15 @@
  *  tables in the database, leveraging the report-specific record types.
  */
 
-use std::str::FromStr;
-
-use carbide_uuid::UuidEmptyStringError;
 use carbide_uuid::machine::MachineId;
 use carbide_uuid::measured_boot::MeasurementReportId;
 use chrono::Utc;
-#[cfg(feature = "cli")]
-use rpc::admin_cli::ToTable;
-use rpc::errors::RpcDataConversionError;
-use rpc::protos::measured_boot::MeasurementReportPb;
 use serde::Serialize;
 
 use super::pcr::PcrRegisterValue;
 use super::records::MeasurementReportValueRecord;
-use crate::{FromGrpc, FromGrpcOpt};
+#[cfg(feature = "cli")]
+use crate::ToTable;
 
 /// MeasurementReport is a composition of a MeasurementReportRecord,
 /// whose attributes are essentially copied directly it, as well as
@@ -58,56 +52,6 @@ impl MeasurementReport {
 impl crate::DisplayName for MeasurementReport {
     fn display_name() -> &'static str {
         "report"
-    }
-}
-
-impl FromGrpc<MeasurementReportPb> for MeasurementReport {}
-
-impl FromGrpcOpt<MeasurementReportPb> for MeasurementReport {}
-
-impl From<MeasurementReport> for MeasurementReportPb {
-    fn from(val: MeasurementReport) -> Self {
-        Self {
-            report_id: Some(val.report_id),
-            machine_id: val.machine_id.to_string(),
-            values: val
-                .values
-                .iter()
-                .map(|value| value.clone().into())
-                .collect(),
-            ts: Some(val.ts.into()),
-        }
-    }
-}
-
-impl TryFrom<MeasurementReportPb> for MeasurementReport {
-    type Error = Box<dyn std::error::Error>;
-
-    fn try_from(msg: MeasurementReportPb) -> Result<Self, Box<dyn std::error::Error>> {
-        if msg.machine_id.is_empty() {
-            return Err(UuidEmptyStringError {}.into());
-        }
-        let values: super::Result<Vec<MeasurementReportValueRecord>> = msg
-            .values
-            .iter()
-            .map(
-                |value| match MeasurementReportValueRecord::try_from(value.clone()) {
-                    Ok(worked) => Ok(worked),
-                    Err(failed) => Err(super::Error::RpcConversion(format!(
-                        "attr conversion failed: {failed}"
-                    ))),
-                },
-            )
-            .collect();
-
-        Ok(Self {
-            report_id: msg
-                .report_id
-                .ok_or(RpcDataConversionError::MissingArgument("report_id"))?,
-            machine_id: MachineId::from_str(&msg.machine_id)?,
-            values: values?,
-            ts: chrono::DateTime::<chrono::Utc>::try_from(msg.ts.unwrap())?,
-        })
     }
 }
