+++ crates/rpc/Cargo.toml
@@ -35,6 +35,7 @@ carbide-http-connector = { path = "../http-connector" }
 carbide-tls = { path = "../tls" }
 carbide-health-report = { path = "../health-report" }
 carbide-libmlx-model = { path = "../libmlx-model" }
+carbide-measured-boot = { path = "../measured-boot", default-features = false }
 carbide-secrets = { path = "../secrets" }
 carbide-network = { path = "../network" }
 carbide-utils = { path = "../utils" }
