+++ crates/rpc/src/lib.rs
@@ -63,6 +63,7 @@ pub use crate::protos::{fmds, health, site_explorer};
 pub mod errors;
 pub mod forge_tls_client;
 pub mod libmlx;
+pub mod measured_boot;
 pub mod network;
 pub mod protos;
 pub mod secrets;
